fix: signUp여부에 따른 dto mapper 구현
