+++ api/src/main/java/com/tago/api/member/application/MemberService.java
@@ -1,5 +1,6 @@
 package com.tago.api.member.application;
 
+import com.tago.api.common.mapper.MemberInfoDtoMapper;
 import com.tago.api.member.dto.response.MemberInfoResponse;
 import com.tago.domain.member.domain.Member;
 import com.tago.domain.member.service.MemberQueryService;
@@ -16,6 +17,6 @@ public class MemberService {
     @Transactional(readOnly = true)
     public MemberInfoResponse getInfo(Long memberId) {
         Member member = memberQueryService.findById(memberId);
-        return MemberInfoResponse.from(member);
+        return MemberInfoDtoMapper.from(member);
     }
 }
